feat(list): change FlatList to FlashList to improve performance
