+++ example/package.json
@@ -9,22 +9,21 @@
     "build:ios": "cd ios && xcodebuild -workspace MessyExample.xcworkspace -scheme MessyExample -configuration Debug -sdk iphonesimulator CC=clang CPLUSPLUS=clang++ LD=clang LDPLUSPLUS=clang++ GCC_OPTIMIZATION_LEVEL=0 GCC_PRECOMPILE_PREFIX_HEADER=YES ASSETCATALOG_COMPILER_OPTIMIZATION=time DEBUG_INFORMATION_FORMAT=dwarf COMPILER_INDEX_STORE_ENABLE=NO"
   },
   "dependencies": {
-    "@candlefinance/faster-image": "^1.5.0",
-    "@gorhom/bottom-sheet": "^4.6.4",
-    "@shopify/flash-list": "^1.6.3",
+    "@gorhom/bottom-sheet": "5.0.0-alpha.11",
+    "@shopify/flash-list": "^1.7.1",
     "@vokhuyet/react-native-messy": "link:../",
     "react": "18.2.0",
     "react-native": "0.73.2",
-    "react-native-gesture-handler": "^2.14.1",
+    "react-native-gesture-handler": "^2.20.0",
     "react-native-image-picker": "^7.1.2",
     "react-native-keyboard-controller": "^1.10.3",
-    "react-native-lightbox-v2": "^0.9.0",
-    "react-native-mmkv-storage": "^0.9.1",
-    "react-native-pager-view": "^6.2.3",
+    "react-native-mmkv-storage": "^0.10.3",
+    "react-native-pager-view": "^6.4.1",
     "react-native-permissions": "^4.1.5",
-    "react-native-reanimated": "^3.6.1",
+    "react-native-reanimated": "^3.15.3",
     "react-native-tab-view": "^3.5.2",
-    "react-native-video": "6.0.0-rc.0"
+    "react-native-turbo-image": "^1.19.1",
+    "react-native-video": "^6.6.3"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
