+++ package.json
@@ -57,16 +57,16 @@
     "dayjs": "^1.11.10",
     "emojilib": "^3.0.11",
     "react-native-parsed-text": "^0.0.22",
-    "react-native-zoom-reanimated": "^1.4.5",
-    "unicode-emoji-json": "^0.4.0"
+    "react-native-zoom-reanimated": "^1.4.8",
+    "unicode-emoji-json": "^0.6.0"
   },
   "devDependencies": {
     "@commitlint/config-conventional": "^17.0.2",
     "@evilmartians/lefthook": "^1.5.0",
-    "@gorhom/bottom-sheet": "^4.6.4",
+    "@gorhom/bottom-sheet": "5.0.0-alpha.11",
     "@react-native/eslint-config": "^0.72.2",
     "@release-it/conventional-changelog": "^5.0.0",
-    "@shopify/flash-list": "^1.6.3",
+    "@shopify/flash-list": "^1.7.1",
     "@types/invariant": "^2.2.37",
     "@types/jest": "^28.1.2",
     "@types/react": "^18.2.61",
@@ -82,13 +82,13 @@
     "react": "18.2.0",
     "react-native": "0.73.2",
     "react-native-builder-bob": "^0.20.0",
-    "react-native-gesture-handler": "^2.14.1",
+    "react-native-gesture-handler": "^2.20.0",
     "react-native-image-picker": "^7.1.2",
-    "react-native-keyboard-controller": "^1.10.3",
-    "react-native-pager-view": "^6.2.3",
-    "react-native-reanimated": "^3.6.1",
+    "react-native-keyboard-controller": "^1.13.4",
+    "react-native-pager-view": "^6.4.1",
+    "react-native-reanimated": "^3.15.3",
     "react-native-tab-view": "^3.5.2",
-    "react-native-video": "6.0.0-rc.0",
+    "react-native-video": "6.6.3",
     "release-it": "^15.0.0",
     "typescript": "^5.0.2"
   },
