+++ src/modules/useMessyListRef.ts
@@ -1,12 +1,16 @@
 import React from 'react';
 import { FlatList } from 'react-native';
 import { BottomSheetFlatListMethods } from '@gorhom/bottom-sheet';
+import type { FlashList } from '@shopify/flash-list';
 
-const listRef = React.createRef<FlatList | BottomSheetFlatListMethods>();
+const listRef = React.createRef<
+  FlatList | BottomSheetFlatListMethods | FlashList<any>
+>();
 
 export const useMessyListRef = () => {
   return listRef;
 };
+
 let timeout: NodeJS.Timeout;
 export const useMessyListAction = () => {
   const scrollToLast = () => {
